updated strlcat and strrchr, part one of test is almost finished

+++ crss/libft/ft_strlcat.c
@@ -14,16 +14,16 @@
 
 size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 {
-	int	i;
-	int	dst_len;
-	int	src_len;
+	size_t	i;
+	size_t	dst_len;
+	size_t	src_len;
 
 	i = 0;
-	dst_len = (int)ft_strlen(dst);
-	src_len = (int)ft_strlen(src);
+	dst_len = ft_strlen(dst);
+	src_len = ft_strlen(src);
 	if (dstsize == 0)
 		return (src_len);
-	while (i < (int)dstsize - 1 && src[i] != '\0')
+	while (i < dstsize - 1 && src[i] != '\0')
 	{
 		dst[dst_len + i] = src[i];
 		i++;
